+++ test/utils/atem_header.h
@@ -3,6 +3,7 @@
 #define ATEM_HEADER_H
 
 #include <stdint.h> // uint8_t, uint16_t
+#include <stddef.h> // size_t
 
 void atem_packet_word_set(uint8_t* packet, uint8_t high, uint8_t low, uint16_t word);
 uint16_t atem_packet_word_get(uint8_t* packet, uint8_t high, uint8_t low);
@@ -18,7 +19,7 @@ void atem_header_flags_isnotset(uint8_t* packet, uint8_t flags);
 
 void atem_header_len_set(uint8_t* packet, uint16_t len);
 uint16_t atem_header_len_get(uint8_t* packet);
-void atem_header_len_get_verify(uint8_t* packet, uint16_t expectedLen);
+void atem_header_len_get_verify(uint8_t* packet, size_t expectedLen);
 
 void atem_header_sessionid_set(uint8_t* packet, uint16_t sessionId);
 uint16_t atem_header_sessionid_get(uint8_t* packet);
